assignment_management/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-widget-tweaks
django-redis==5.2.0
django-select2==8.1.1
django-widget-tweaks==1.4.12
python-dateutil==2.8.2
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
# opencv-python
# opencv-contrib-python